MU::Config: tear Ref and Tail out into their own files; make some class methods public again, private doesn't fly
