+++ modules/mu/config.rb
@@ -18,6 +18,8 @@ require 'erb'
 require 'pp'
 require 'json-schema'
 require 'net/http'
+require 'mu/config/tail'
+require 'mu/config/ref'
 autoload :GraphViz, 'graphviz'
 autoload :ChronicDuration, 'chronic_duration'
 
@@ -241,488 +243,6 @@ module MU
     # @return [Hash]
     def self.stripConfig(config)
       MU::Config.manxify(Marshal.load(Marshal.dump(MU.structToHash(config.dup))), remove_runtime_keys: true)
-    end
-
-    # A wrapper class for resources to refer to other resources, whether they
-    # be a sibling object in the current deploy, an object in another deploy,
-    # or a plain cloud id from outside of Mu.
-    class Ref
-      attr_reader :name 
-      attr_reader :type 
-      attr_reader :cloud 
-      attr_reader :deploy_id 
-      attr_reader :region 
-      attr_reader :credentials 
-      attr_reader :habitat
-      attr_reader :mommacat
-      attr_reader :tag_key 
-      attr_reader :tag_value
-      attr_reader :obj 
-
-      @@refs = []
-      @@ref_semaphore = Mutex.new
-
-      # Little bit of a factory pattern... given a hash of options for a {MU::Config::Ref} objects, first see if we have an existing one that matches our more immutable attributes (+cloud+, +id+, etc). If we do, return that. If we do not, create one, add that to our inventory, and return that instead.
-      # @param cfg [Hash]: 
-      # @return [MU::Config::Ref]
-      def self.get(cfg)
-        return cfg if cfg.is_a?(MU::Config::Ref)
-        checkfields = cfg.keys.map { |k| k.to_sym }
-        required = [:id, :type]
-
-        @@ref_semaphore.synchronize {
-          @@refs.each { |ref|
-            saw_mismatch = false
-            saw_match = false
-            needed_values = []
-            checkfields.each { |field|
-              next if !cfg[field]
-              ext_value = ref.instance_variable_get("@#{field.to_s}".to_sym)
-              if !ext_value
-                needed_values << field
-                next
-              end
-              if cfg[field] != ext_value
-                saw_mismatch = true
-              elsif required.include?(field) and cfg[field] == ext_value
-                saw_match = true
-              end
-            }
-            if saw_match and !saw_mismatch
-              # populate empty fields we got from this request
-              if needed_values.size > 0
-                newref = ref.dup
-                needed_values.each { |field|
-                  newref.instance_variable_set("@#{field.to_s}".to_sym, cfg[field])
-                  if !newref.respond_to?(field)
-                    newref.singleton_class.instance_eval { attr_reader field.to_sym }
-                  end
-                }
-                @@refs << newref
-                return newref
-              else
-                return ref
-              end
-            end
-          }
-
-        }
-
-        # if we get here, there was no match
-        newref = MU::Config::Ref.new(cfg)
-        @@ref_semaphore.synchronize {
-          @@refs << newref
-          return newref
-        }
-      end
-
-      # A way of dynamically defining +attr_reader+ without leaking memory
-      def self.define_reader(name)
-        define_method(name) {
-          instance_variable_get("@#{name.to_s}")
-        }
-      end
-
-      # @param cfg [Hash]: A Basket of Kittens configuration hash containing
-      # lookup information for a cloud object
-      def initialize(cfg)
-        cfg.keys.each { |field|
-          next if field == "tag"
-          if !cfg[field].nil?
-            self.instance_variable_set("@#{field}".to_sym, cfg[field])
-          elsif !cfg[field.to_sym].nil?
-            self.instance_variable_set("@#{field.to_s}".to_sym, cfg[field.to_sym])
-          end
-          MU::Config::Ref.define_reader(field)
-        }
-        if cfg['tag'] and cfg['tag']['key'] and
-           !cfg['tag']['key'].empty? and cfg['tag']['value']
-          @tag_key = cfg['tag']['key']
-          @tag_value = cfg['tag']['value']
-        end
-
-        if @deploy_id and !@mommacat
-          @mommacat = MU::MommaCat.getLitter(@deploy_id, set_context_to_me: false)
-        elsif @mommacat and !@deploy_id
-          @deploy_id = @mommacat.deploy_id
-        end
-
-        kitten(shallow: true) if @mommacat # try to populate the actual cloud object for this
-      end
-
-      # Comparison operator
-      def <=>(other)
-        return 1 if other.nil?
-        self.to_s <=> other.to_s
-      end
-
-      # Base configuration schema for declared kittens referencing other cloud objects. This is essentially a set of filters that we're going to pass to {MU::MommaCat.findStray}.
-      # @param aliases [Array<Hash>]: Key => value mappings to set backwards-compatibility aliases for attributes, such as the ubiquitous +vpc_id+ (+vpc_id+ => +id+).
-      # @return [Hash]
-      def self.schema(aliases = [], type: nil, parent_obj: nil, desc: nil, omit_fields: [])
-        parent_obj ||= caller[1].gsub(/.*?\/([^\.\/]+)\.rb:.*/, '\1')
-        desc ||= "Reference a #{type ? "'#{type}' resource" : "resource" } from this #{parent_obj ? "'#{parent_obj}'" : "" } resource"
-        schema = {
-          "type" => "object",
-          "#MU_REFERENCE" => true,
-          "minProperties" => 1,
-          "description" => desc,
-          "properties" => {
-            "id" => {
-              "type" => "string",
-              "description" => "Cloud identifier of a resource we want to reference, typically used when leveraging resources not managed by MU"
-            },
-            "name" => {
-              "type" => "string",
-              "description" => "The short (internal Mu) name of a resource we're attempting to reference. Typically used when referring to a sibling resource elsewhere in the same deploy, or in another known Mu deploy in conjunction with +deploy_id+."
-            },
-            "type" => {
-              "type" => "string",
-              "description" => "The resource type we're attempting to reference.",
-              "enum" => MU::Cloud.resource_types.values.map { |t| t[:cfg_plural] }
-            },
-            "deploy_id" => {
-              "type" => "string",
-              "description" => "Our target resource should be found in this Mu deploy."
-            },
-            "credentials" => MU::Config.credentials_primitive,
-            "region" => MU::Config.region_primitive,
-            "cloud" => MU::Config.cloud_primitive,
-            "tag" => {
-              "type" => "object",
-              "description" => "If the target resource supports tagging and our resource implementations +find+ method supports it, we can attempt to locate it by tag.",
-              "properties" => {
-                "key" => {
-                  "type" => "string",
-                  "description" => "The tag or label key to search against"
-                },
-                "value" => {
-                  "type" => "string",
-                  "description" => "The tag or label value to match"
-                }
-              }
-            }
-          }
-        }
-        if !["folders", "habitats"].include?(type)
-          schema["properties"]["habitat"] = MU::Config::Habitat.reference
-        end
-
-        if omit_fields
-          omit_fields.each { |f|
-            schema["properties"].delete(f)
-          }
-        end
-
-        if !type.nil?
-          schema["required"] = ["type"]
-          schema["properties"]["type"]["default"] = type
-          schema["properties"]["type"]["enum"] = [type]
-        end
-
-        aliases.each { |a|
-          a.each_pair { |k, v|
-            if schema["properties"][v]
-              schema["properties"][k] = schema["properties"][v].dup
-              schema["properties"][k]["description"] = "Alias for <tt>#{v}</tt>"
-            else
-              MU.log "Reference schema alias #{k} wants to alias #{v}, but no such attribute exists", MU::WARN, details: caller[4]
-            end
-          }
-        }
-
-        schema
-      end
-
-      # Decompose into a plain-jane {MU::Config::BasketOfKittens} hash fragment,
-      # of the sort that would have been used to declare this reference in the
-      # first place.
-      def to_h
-        me = { }
-
-        self.instance_variables.each { |var|
-          next if [:@obj, :@mommacat, :@tag_key, :@tag_value].include?(var)
-          val = self.instance_variable_get(var)
-          next if val.nil?
-          val = val.to_h if val.is_a?(MU::Config::Ref)
-          me[var.to_s.sub(/^@/, '')] = val
-        }
-        if @tag_key and !@tag_key.empty?
-          me['tag'] = {
-            'key' => @tag_key,
-            'value' => @tag_value
-          }
-        end
-        me
-      end
-
-      # Getter for the #{id} instance variable that attempts to populate it if
-      # it's not set.
-      # @return [String,nil]
-      def id
-        return @id if @id
-        kitten # if it's not defined, attempt to define it
-        @id
-      end
-
-      # Alias for {id}
-      # @return [String,nil]
-      def cloud_id
-        id
-      end
-
-      # Return a {MU::Cloud} object for this reference. This is only meant to be
-      # called in a live deploy, which is to say that if called during initial
-      # configuration parsing, results may be incorrect.
-      # @param mommacat [MU::MommaCat]: A deploy object which will be searched for the referenced resource if provided, before restoring to broader, less efficient searches.
-      def kitten(mommacat = @mommacat, shallow: false)
-        return nil if !@cloud or !@type
-
-        if @obj
-          @deploy_id ||= @obj.deploy_id
-          @id ||= @obj.cloud_id
-          @name ||= @obj.config['name']
-          return @obj
-        end
-
-        if mommacat
-          @obj = mommacat.findLitterMate(type: @type, name: @name, cloud_id: @id, credentials: @credentials, debug: false)
-          if @obj # initialize missing attributes, if we can
-            @id ||= @obj.cloud_id
-            @mommacat ||= mommacat
-            @obj.intoDeploy(@mommacat) # make real sure these are set
-            @deploy_id ||= mommacat.deploy_id
-            if !@name
-              if @obj.config and @obj.config['name']
-                @name = @obj.config['name']
-              elsif @obj.mu_name
-if @type == "folders"
-MU.log "would assign name '#{@obj.mu_name}' in ref to this folder if I were feeling aggressive", MU::WARN, details: self.to_h
-end
-#                @name = @obj.mu_name
-              end
-            end
-            return @obj
-          else
-#            MU.log "Failed to find a live '#{@type.to_s}' object named #{@name}#{@id ? " (#{@id})" : "" }#{ @habitat ? " in habitat #{@habitat}" : "" }", MU::WARN, details: self
-          end
-        end
-
-        if !@obj and !(@cloud == "Google" and @id and @type == "users" and MU::Cloud::Google::User.cannedServiceAcctName?(@id)) and !shallow
-
-          begin
-            hab_arg = if @habitat.nil?
-              [nil]
-            elsif @habitat.is_a?(MU::Config::Ref)
-              [@habitat.id]
-            elsif @habitat.is_a?(Hash)
-              [@habitat["id"]]
-            else
-              [@habitat.to_s]
-            end
-
-            found = MU::MommaCat.findStray(
-              @cloud,
-              @type,
-              name: @name,
-              cloud_id: @id,
-              deploy_id: @deploy_id,
-              region: @region,
-              habitats: hab_arg,
-              credentials: @credentials,
-              dummy_ok: (["habitats", "folders", "users", "groups", "vpcs"].include?(@type))
-            )
-            @obj ||= found.first if found
-          rescue ThreadError => e
-            # Sometimes MommaCat calls us in a potential deadlock situation;
-            # don't be the cause of a fatal error if so, we don't need this
-            # object that badly.
-            raise e if !e.message.match(/recursive locking/)
-rescue SystemExit
-# XXX this is temporary, to cope with some debug stuff that's in findStray
-# for the nonce
-return
-          end
-        end
-
-        if @obj
-          @deploy_id ||= @obj.deploy_id
-          @id ||= @obj.cloud_id
-          @name ||= @obj.config['name']
-        end
-
-        @obj
-      end
-
-    end
-
-    # A wrapper for config leaves that came from ERB parameters instead of raw
-    # YAML or JSON. Will behave like a string for things that expect that
-    # sort of thing. Code that needs to know that this leaf was the result of
-    # a parameter will be able to tell by the object class being something
-    # other than a plain string, array, or hash.
-    class Tail
-      @value = nil
-      @name = nil
-      @prettyname = nil
-      @description = nil
-      @prefix = ""
-      @suffix = ""
-      @is_list_element = false
-      @pseudo = false
-      @runtimecode = nil
-      @valid_values = []
-      @index = 0
-      attr_reader :description
-      attr_reader :pseudo
-      attr_reader :index
-      attr_reader :value
-      attr_reader :runtimecode
-      attr_reader :valid_values
-      attr_reader :is_list_element
-
-      def initialize(name, value, prettyname = nil, cloudtype = "String", valid_values = [], description = "", is_list_element = false, prefix: "", suffix: "", pseudo: false, runtimecode: nil, index: 0)
-        @name = name
-        @bindings = {}
-        @value = value
-        @valid_values = valid_values
-        @pseudo = pseudo
-        @index = index
-        @runtimecode = runtimecode
-        @cloudtype = cloudtype
-        @is_list_element = is_list_element
-        @description ||= 
-          if !description.nil?
-            description
-          else
-            ""
-          end
-        @prettyname ||= 
-          if !prettyname.nil?
-            prettyname
-          else
-            @name.capitalize.gsub(/[^a-z0-9]/i, "")
-          end
-        @prefix = prefix if !prefix.nil?
-        @suffix = suffix if !suffix.nil?
-      end
- 
-      # Return the parameter name of this Tail
-      def getName
-        @name
-      end
-      # Return the platform-specific cloud type of this Tail
-      def getCloudType
-        @cloudtype
-      end
-      # Return the human-friendly name of this Tail
-      def getPrettyName
-        @prettyname
-      end
-      # Walk like a String
-      def to_s
-        @prefix.to_s+@value.to_s+@suffix.to_s
-      end
-      # Quack like a String
-      def to_str
-        to_s
-      end
-      # Upcase like a String
-      def upcase
-        to_s.upcase
-      end
-      # Downcase like a String
-      def downcase
-        to_s.downcase
-      end
-      # Check for emptiness like a String
-      def empty?
-        to_s.empty?
-      end
-      # Match like a String
-      def match(*args)
-        to_s.match(*args)
-      end
-      # Check for equality like a String
-      def ==(o)
-        (o.class == self.class or o.class == "String") && o.to_s == to_s
-      end
-      # Concatenate like a string
-      def +(o)
-        return to_s if o.nil?
-        to_s + o.to_s
-      end
-      # Perform global substitutions like a String
-      def gsub(*args)
-        to_s.gsub(*args)
-      end
-    end
-
-    # Wrapper method for creating a {MU::Config::Tail} object as a reference to
-    # a parameter that's valid in the loaded configuration.
-    # @param param [<String>]: The name of the parameter to which this should be tied.
-    # @param value [<String>]: The value of the parameter to return when asked
-    # @param prettyname [<String>]: A human-friendly parameter name to be used when generating CloudFormation templates and the like
-    # @param cloudtype [<String>]: A platform-specific identifier used by cloud layers to identify a parameter's type, e.g. AWS::EC2::VPC::Id
-    # @param valid_values [Array<String>]: A list of acceptable String values for the given parameter.
-    # @param description [<String>]: A long-form description of what the parameter does.
-    # @param list_of [<String>]: Indicates that the value should be treated as a member of a list (array) by the cloud layer.
-    # @param prefix [<String>]: A static String that should be prefixed to the stored value when queried
-    # @param suffix [<String>]: A static String that should be appended to the stored value when queried
-    # @param pseudo [<Boolean>]: This is a pseudo-parameter, automatically provided, and not available as user input.
-    # @param runtimecode [<String>]: Actual code to allow the cloud layer to interpret literally in its own idiom, e.g. '"Ref" : "AWS::StackName"' for CloudFormation
-    def getTail(param, value: nil, prettyname: nil, cloudtype: "String", valid_values: [], description: nil, list_of: nil, prefix: "", suffix: "", pseudo: false, runtimecode: nil)
-      if value.nil?
-        if @@parameters.nil? or !@@parameters.has_key?(param)
-          MU.log "Parameter '#{param}' (#{param.class.name}) referenced in config but not provided (#{caller[0]})", MU::DEBUG, details: @@parameters
-          return nil
-#          raise DeployParamError
-        else
-          value = @@parameters[param]
-        end
-      end
-      if !prettyname.nil?
-        prettyname.gsub!(/[^a-z0-9]/i, "") # comply with CloudFormation restrictions
-      end
-      if value.is_a?(MU::Config::Tail)
-        MU.log "Parameter #{param} is using a nested parameter as a value. This rarely works, depending on the target cloud. YMMV.", MU::WARN
-        tail = MU::Config::Tail.new(param, value, prettyname, cloudtype, valid_values, description, prefix: prefix, suffix: suffix, pseudo: pseudo, runtimecode: runtimecode)
-      elsif !list_of.nil? or (@@tails.has_key?(param) and @@tails[param].is_a?(Array))
-        tail = []
-        count = 0
-        value.split(/\s*,\s*/).each { |subval|
-          if @@tails.has_key?(param) and !@@tails[param][count].nil?
-            subval = @@tails[param][count].values.first.to_s if subval.nil?
-            list_of = @@tails[param][count].values.first.getName if list_of.nil?
-            prettyname = @@tails[param][count].values.first.getPrettyName if prettyname.nil?
-            description = @@tails[param][count].values.first.description if description.nil?
-            valid_values = @@tails[param][count].values.first.valid_values if valid_values.nil? or valid_values.empty?
-            cloudtype = @@tails[param][count].values.first.getCloudType if @@tails[param][count].values.first.getCloudType != "String"
-          end
-          prettyname = param.capitalize if prettyname.nil?
-          tail << { list_of => MU::Config::Tail.new(list_of, subval, prettyname, cloudtype, valid_values, description, true, pseudo: pseudo, index: count) }
-          count = count + 1
-        }
-      else
-        if @@tails.has_key?(param)
-          pseudo = @@tails[param].pseudo
-          value = @@tails[param].to_s if value.nil?
-          prettyname = @@tails[param].getPrettyName if prettyname.nil?
-          description = @@tails[param].description if description.nil?
-          valid_values = @@tails[param].valid_values if valid_values.nil? or valid_values.empty?
-          cloudtype = @@tails[param].getCloudType if @@tails[param].getCloudType != "String"
-        end
-        tail = MU::Config::Tail.new(param, value, prettyname, cloudtype, valid_values, description, prefix: prefix, suffix: suffix, pseudo: pseudo, runtimecode: runtimecode)
-      end
-
-      if valid_values and valid_values.size > 0 and value
-        if !valid_values.include?(value)
-          raise DeployParamError, "Invalid parameter value '#{value}' supplied for '#{param}'"
-        end
-      end
-      @@tails[param] = tail
-
-      tail
     end
 
     # Load up our YAML or JSON and parse it through ERB, optionally substituting
@@ -2381,6 +1901,65 @@ $CONFIGURABLES
       @@bindings
     end
 
+    # Verify that a server or server_pool has a valid AD config referencing
+    # valid Vaults for credentials.
+    # @param server [MU::Cloud::Server]
+    def self.check_vault_refs(server)
+      ok = true
+      server['vault_access'] = [] if server['vault_access'].nil?
+      server['groomer'] ||= self.defaultGroomer
+      groomclass = MU::Groomer.loadGroomer(server['groomer'])
+
+      begin
+        if !server['active_directory'].nil?
+          ["domain_admin_vault", "domain_join_vault"].each { |vault_class|
+            server['vault_access'] << {
+                "vault" => server['active_directory'][vault_class]['vault'],
+                "item" => server['active_directory'][vault_class]['item']
+            }
+            item = groomclass.getSecret(
+                vault: server['active_directory'][vault_class]['vault'],
+                item: server['active_directory'][vault_class]['item'],
+            )
+            ["username_field", "password_field"].each { |field|
+              if !item.has_key?(server['active_directory'][vault_class][field])
+                ok = false
+                MU.log "I don't see a value named #{field} in Chef Vault #{server['active_directory'][vault_class]['vault']}:#{server['active_directory'][vault_class]['item']}", MU::ERR
+              end
+            }
+          }
+        end
+
+        if !server['windows_auth_vault'].nil?
+          server['use_cloud_provider_windows_password'] = false
+
+          server['vault_access'] << {
+              "vault" => server['windows_auth_vault']['vault'],
+              "item" => server['windows_auth_vault']['item']
+          }
+          item = groomclass.getSecret(
+            vault: server['windows_auth_vault']['vault'],
+            item: server['windows_auth_vault']['item']
+          )
+          ["password_field", "ec2config_password_field", "sshd_password_field"].each { |field|
+            if !item.has_key?(server['windows_auth_vault'][field])
+              MU.log "No value named #{field} in Chef Vault #{server['windows_auth_vault']['vault']}:#{server['windows_auth_vault']['item']}, will use a generated password.", MU::NOTICE
+              server['windows_auth_vault'].delete(field)
+            end
+          }
+        end
+        # Check all of the non-special ones while we're at it
+        server['vault_access'].each { |v|
+          next if v['vault'] == "splunk" and v['item'] == "admin_user"
+          item = groomclass.getSecret(vault: v['vault'], item: v['item'])
+        }
+      rescue MuError
+        MU.log "Can't load a Chef Vault I was configured to use. Does it exist?", MU::ERR
+        ok = false
+      end
+      return ok
+    end
+
     private
 
     # (see #include)
@@ -2456,65 +2035,7 @@ $CONFIGURABLES
       return conf_chunk
     end
 
-    # Verify that a server or server_pool has a valid AD config referencing
-    # valid Vaults for credentials.
-    def self.check_vault_refs(server)
-      ok = true
-      server['vault_access'] = [] if server['vault_access'].nil?
-      server['groomer'] ||= self.defaultGroomer
-      groomclass = MU::Groomer.loadGroomer(server['groomer'])
-
-      begin
-        if !server['active_directory'].nil?
-          ["domain_admin_vault", "domain_join_vault"].each { |vault_class|
-            server['vault_access'] << {
-                "vault" => server['active_directory'][vault_class]['vault'],
-                "item" => server['active_directory'][vault_class]['item']
-            }
-            item = groomclass.getSecret(
-                vault: server['active_directory'][vault_class]['vault'],
-                item: server['active_directory'][vault_class]['item'],
-            )
-            ["username_field", "password_field"].each { |field|
-              if !item.has_key?(server['active_directory'][vault_class][field])
-                ok = false
-                MU.log "I don't see a value named #{field} in Chef Vault #{server['active_directory'][vault_class]['vault']}:#{server['active_directory'][vault_class]['item']}", MU::ERR
-              end
-            }
-          }
-        end
-
-        if !server['windows_auth_vault'].nil?
-          server['use_cloud_provider_windows_password'] = false
-
-          server['vault_access'] << {
-              "vault" => server['windows_auth_vault']['vault'],
-              "item" => server['windows_auth_vault']['item']
-          }
-          item = groomclass.getSecret(
-            vault: server['windows_auth_vault']['vault'],
-            item: server['windows_auth_vault']['item']
-          )
-          ["password_field", "ec2config_password_field", "sshd_password_field"].each { |field|
-            if !item.has_key?(server['windows_auth_vault'][field])
-              MU.log "No value named #{field} in Chef Vault #{server['windows_auth_vault']['vault']}:#{server['windows_auth_vault']['item']}, will use a generated password.", MU::NOTICE
-              server['windows_auth_vault'].delete(field)
-            end
-          }
-        end
-        # Check all of the non-special ones while we're at it
-        server['vault_access'].each { |v|
-          next if v['vault'] == "splunk" and v['item'] == "admin_user"
-          item = groomclass.getSecret(vault: v['vault'], item: v['item'])
-        }
-      rescue MuError
-        MU.log "Can't load a Chef Vault I was configured to use. Does it exist?", MU::ERR
-        ok = false
-      end
-      return ok
-    end
-    private_class_method :check_vault_refs
-    
+ 
     # Given a bare hash describing a resource, insert default values which can
     # be inherited from its parent or from the root of the BoK.
     # @param kitten [Hash]: A resource descriptor
